Failing to refactor fractal generator
